backend/analytics-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
alembic
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dateutil
requests